# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Library Book Tracker System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /borrow:
    post:
      operationId: borrowBook
      tags: [borrow]
      summary: Borrow a book
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BorrowBody"
      responses:
        "201":
          description: Book borrowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BorrowRecord"
    get:
      operationId: getBorrowHistory
      tags: [borrow]
      summary: Get borrow history for current user
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Borrow history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BorrowHistoryResponse"
  /borrow/active:
    get:
      operationId: getActiveBorrows
      tags: [borrow]
      summary: Get all active borrows (admin/librarian)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active borrows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BorrowRecord"
  /borrow/return:
    post:
      operationId: returnBook
      tags: [borrow]
      summary: Return a borrowed book
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnBody"
      responses:
        "200":
          description: Book returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BorrowRecord"
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get current user's wishlist
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add a book to wishlist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistBody"
      responses:
        "201":
          description: Added to wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
  /wishlist/{bookId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove a book from wishlist
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, librarian, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        subject:
          type: string
        branch:
          type: string
          enum: [CSE, IT, Civil, Mechanical, Electrical]
        availability:
          type: boolean
        section:
          type: string
        rackNumber:
          type: string
        rowNumber:
          type: string
        shelfNumber:
          type: string
        isbn:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - author
        - subject
        - branch
        - availability
        - section
        - rackNumber
        - rowNumber
        - shelfNumber
...
    BorrowBody:
      type: object
      properties:
        bookId:
          type: integer
        dueDays:
          type: integer
          default: 14
      required:
        - bookId
    ReturnBody:
      type: object
      properties:
        borrowId:
          type: integer
      required:
        - borrowId
    BorrowRecord:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        bookId:
          type: integer
        book:
          $ref: "#/components/schemas/Book"
        user:
          $ref: "#/components/schemas/User"
        borrowDate:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        returnDate:
          type: string
          format: date-time
          nullable: true
        isOverdue:
          type: boolean
        fineDays:
          type: integer
          nullable: true
      required:
        - id
        - userId
        - bookId
        - borrowDate
        - dueDate
        - isOverdue
    BorrowHistoryResponse:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/BorrowRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - records
        - total
        - page
        - limit
        - totalPages
    WishlistBody:
      type: object
      properties:
        bookId:
          type: integer
      required:
        - bookId
    WishlistItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        bookId:
          type: integer
        book:
          $ref: "#/components/schemas/Book"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - bookId
...